feat(client): generate index.ts file for models, schemas, and services
