+++ .changeset/orange-keys-fry.md
@@ -1,0 +1,5 @@
+---
+"@hey-api/openapi-ts": minor
+---
+
+Add `index.ts` file to models, schemas, and services
