+++ rollup.config.ts
@@ -42,9 +42,6 @@ export function handlebarsPlugin(): Plugin {
                         escapeDescription: true,
                         escapeNewline: true,
                         exactArray: true,
-                        exportsModels: true,
-                        exportsSchemas: true,
-                        exportsServices: true,
                         ifdef: true,
                         ifOperationDataOptional: true,
                         intersection: true,
