+++ src/utils/write/models.ts
@@ -3,6 +3,7 @@ import path from 'node:path';
 
 import type { Client } from '../../types/client';
 import type { Config } from '../../types/config';
+import { enumName } from '../enum';
 import type { Templates } from '../handlebars';
 
 /**
@@ -18,6 +19,7 @@ export const writeClientModels = async (
     outputPath: string,
     config: Config
 ): Promise<void> => {
+    // Generate a file for each model.
     for (const model of client.models) {
         const file = path.resolve(outputPath, `${model.name}.ts`);
         const templateResult = templates.exports.model({
@@ -26,4 +28,26 @@ export const writeClientModels = async (
         });
         await writeFileSync(file, templateResult);
     }
+    // Generate an index.ts file exporting all models from each file generated above.
+    const file = path.resolve(outputPath, 'index.ts');
+    const content = exportsModels(config, client);
+    await writeFileSync(file, content);
+};
+
+const exportsModels = (config: Config, client: Client) => {
+    const path = './';
+    const output = client.models.map(model => {
+        const importedModel = config.postfixModels
+            ? `${model.name} as ${model.name + config.postfixModels}`
+            : model.name;
+        let result = [`export type { ${importedModel} } from '${path + model.name}';`];
+        if (config.enums && (model.enum.length || model.enums.length)) {
+            const names = model.enums.map(enumerator => enumerator.name).filter(Boolean);
+            const enumExports = names.length ? names : [model.name];
+            const enumExportsString = enumExports.map(name => enumName(name)).join(', ');
+            result = [...result, `export { ${enumExportsString} } from '${path + model.name}';`];
+        }
+        return result.join('\n');
+    });
+    return output.join('\n');
 };
