+++ src/utils/write/schemas.ts
@@ -4,6 +4,7 @@ import path from 'node:path';
 import type { Client } from '../../types/client';
 import type { Config } from '../../types/config';
 import type { Templates } from '../handlebars';
+import { sortByName } from '../sort';
 
 /**
  * Generate Schemas using the Handlebar template and write to disk.
@@ -18,6 +19,7 @@ export const writeClientSchemas = async (
     outputPath: string,
     config: Config
 ): Promise<void> => {
+    // Generate file for each models schema.
     for (const model of client.models) {
         const file = path.resolve(outputPath, `$${model.name}.ts`);
         const templateResult = templates.exports.schema({
@@ -26,4 +28,8 @@ export const writeClientSchemas = async (
         });
         await writeFileSync(file, templateResult);
     }
+    // Generate index file exporting all generated schema files.
+    const file = path.resolve(outputPath, 'index.ts');
+    const content = sortByName(client.models).map(model => `export { $${model.name} } from './$${model.name}';`);
+    await writeFileSync(file, content.join('\n'));
 };
