+++ src/utils/write/services.ts
@@ -4,6 +4,7 @@ import path from 'node:path';
 import type { Client } from '../../types/client';
 import type { Config } from '../../types/config';
 import type { Templates } from '../handlebars';
+import { sortByName } from '../sort';
 
 /**
  * Generate Services using the Handlebar template and write to disk.
@@ -18,6 +19,7 @@ export const writeClientServices = async (
     outputPath: string,
     config: Config
 ): Promise<void> => {
+    // Generate file for each service.
     for (const service of client.services) {
         const file = path.resolve(outputPath, `${service.name}${config.postfixServices}.ts`);
         const templateResult = templates.exports.service({
@@ -26,4 +28,11 @@ export const writeClientServices = async (
         });
         await writeFileSync(file, templateResult);
     }
+    // Generate index file exporting all generated service files.
+    const file = path.resolve(outputPath, 'index.ts');
+    const content = sortByName(client.services).map(
+        service =>
+            `export { ${service.name}${config.postfixServices} } from './${service.name}${config.postfixServices}'`
+    );
+    await writeFileSync(file, content.join('\n'));
 };
